requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Pillow==9.5.0
numpy==1.24.2
gunicorn==20.1.0
